updated code

+++ TP2/code/sequential_with_optimization.h
@@ -1,7 +1,7 @@
 
 #ifndef SEQUENTIAL_WITH_OPTIMIZATION_H
 #define SEQUENTIAL_WITH_OPTIMIZATION_H
-int add_to_array(long number, int div,  int res, long * output);
+int add_to_array(long number, long div,  int res, long * output);
 int sequential_optimization_loop(char *input_string, int size, long * output);
 void sequential_optimization(char *input_string);
 #endif
